+++ library/cpp/coroutine/engine/stack/stack_pool.h
@@ -21,16 +21,16 @@ namespace NCoro::NStack {
             char* Aligned = nullptr; // points to aligned memory, which includes space for first page guard
         };
     public:
-        TPool(uint64_t stackSize, const TPoolAllocatorSettings& settings, const TGuard& guard);
+        TPool(size_t stackSize, const TPoolAllocatorSettings& settings, const TGuard& guard);
         TPool(TPool&& other) noexcept;
         ~TPool();
 
         NDetails::TStack AllocStack(const char* name);
         void FreeStack(NDetails::TStack& stack);
 
-        uint64_t GetReleasedSize() const noexcept;
-        uint64_t GetFullSize() const noexcept;
-        uint64_t GetNumOfAllocated() const noexcept { return NumOfAllocated_; }
+        size_t GetReleasedSize() const noexcept;
+        size_t GetFullSize() const noexcept;
+        size_t GetNumOfAllocated() const noexcept { return NumOfAllocated_; }
 
     private:
         void AllocNewMemoryChunk();
@@ -39,14 +39,14 @@ namespace NCoro::NStack {
         NDetails::TStack AllocNewStack(const char* name);
 
     private:
-        const uint64_t StackSize_ = 0;
-        uint64_t RssPagesToKeep_ = 0;
+        const size_t StackSize_ = 0;
+        size_t RssPagesToKeep_ = 0;
         const TGuard& Guard_;
         TVector<TMemory> Memory_; // memory chunks
         THolder<TStorage> Storage_;
         char* NextToAlloc_ = nullptr; // points to next available stack in the last memory chunk
-        const uint64_t ChunkSize_ = 0;
-        uint64_t NumOfAllocated_ = 0;
+        const size_t ChunkSize_ = 0;
+        size_t NumOfAllocated_ = 0;
     };
 
 }
