+++ library/cpp/coroutine/engine/stack/stack_utils.cpp
@@ -1,6 +1,5 @@
 #include "stack_utils.h"
 
-#include <contrib/libs/linux-headers/asm-generic/errno-base.h>
 #include <util/generic/scope.h>
 #include <util/system/yassert.h>
 
@@ -16,7 +15,7 @@
 namespace NCoro::NStack {
 
 #ifdef _linux_
-    bool GetAlignedMemory(uint64_t sizeInPages, char*& rawPtr, char*& alignedPtr) noexcept {
+    bool GetAlignedMemory(size_t sizeInPages, char*& rawPtr, char*& alignedPtr) noexcept {
         Y_ASSERT(sizeInPages);
 
         void* ptr = nullptr;
@@ -25,18 +24,18 @@ namespace NCoro::NStack {
         return rawPtr && alignedPtr && !error;
     }
 #else
-    bool GetAlignedMemory(uint64_t sizeInPages, char*& rawPtr, char*& alignedPtr) noexcept {
+    bool GetAlignedMemory(size_t sizeInPages, char*& rawPtr, char*& alignedPtr) noexcept {
         Y_ASSERT(sizeInPages);
 
         rawPtr = (char*) malloc((sizeInPages + 1) * PageSize); // +1 in case result would be unaligned
-        alignedPtr = (char*)( ((uint64_t)rawPtr + PageSize - 1) & ~PageSizeMask);
+        alignedPtr = (char*)( ((size_t)rawPtr + PageSize - 1) & ~PageSizeMask);
         return rawPtr && alignedPtr;
     }
 #endif
 
 #ifdef _linux_
-    void ReleaseRss(char* alignedPtr, uint64_t numOfPages) noexcept {
-        Y_VERIFY( !((uint64_t)alignedPtr & PageSizeMask), "Not aligned pointer to release RSS memory");
+    void ReleaseRss(char* alignedPtr, size_t numOfPages) noexcept {
+        Y_VERIFY( !((size_t)alignedPtr & PageSizeMask), "Not aligned pointer to release RSS memory");
         if (!numOfPages) {
             return;
         }
@@ -45,16 +44,16 @@ namespace NCoro::NStack {
         }
     }
 #else
-    void ReleaseRss(char*, uint64_t) noexcept {
+    void ReleaseRss(char*, size_t) noexcept {
     }
 #endif
 
 #ifdef _linux_
-    uint64_t CountMapped(char* alignedPtr, uint64_t numOfPages) noexcept {
-        Y_VERIFY( !((uint64_t)alignedPtr & PageSizeMask) );
+    size_t CountMapped(char* alignedPtr, size_t numOfPages) noexcept {
+        Y_VERIFY( !((size_t)alignedPtr & PageSizeMask) );
         Y_ASSERT(numOfPages);
 
-        uint64_t result = 0;
+        size_t result = 0;
         unsigned char* mappedPages = (unsigned char*) calloc(numOfPages, numOfPages);
         Y_VERIFY(mappedPages);
         Y_DEFER {
@@ -62,7 +61,7 @@ namespace NCoro::NStack {
         };
 
         if (!mincore((void*)alignedPtr, numOfPages * PageSize, mappedPages)) {
-            for (uint64_t i = 0; i < numOfPages; ++i) {
+            for (size_t i = 0; i < numOfPages; ++i) {
                 if (mappedPages[i] & 1) {
                     ++result;
                 }
@@ -76,7 +75,7 @@ namespace NCoro::NStack {
     }
 
 #else
-    uint64_t CountMapped(char*, uint64_t) noexcept {
+    size_t CountMapped(char*, size_t) noexcept {
         return 0; // stub for Windows tests
     }
 #endif
