+++ library/cpp/coroutine/engine/stack/ut/stack_utils_ut.cpp
@@ -9,11 +9,11 @@ namespace NCoro::NStack::Tests {
 
     TEST(StackUtilsTest, Allocation) {
         char *rawPtr, *alignedPtr = nullptr;
-        for (uint64_t i : {1, 2, 3, 4, 11}) {
+        for (size_t i : {1, 2, 3, 4, 11}) {
             EXPECT_TRUE(GetAlignedMemory(i, rawPtr, alignedPtr));
             EXPECT_TRUE(rawPtr);
             EXPECT_TRUE(alignedPtr);
-            EXPECT_FALSE((uint64_t)alignedPtr & PageSizeMask);
+            EXPECT_FALSE((size_t)alignedPtr & PageSizeMask);
             free(rawPtr);
         }
     }
@@ -22,11 +22,11 @@ namespace NCoro::NStack::Tests {
 
     TEST(StackUtilsTest, RssReleaseOnePage) {
         char *rawPtr, *alignedPtr = nullptr;
-        for (uint64_t i : {1, 2, 8}) {
+        for (size_t i : {1, 2, 8}) {
             EXPECT_TRUE(GetAlignedMemory(i, rawPtr, alignedPtr));
             EXPECT_TRUE(rawPtr);
             EXPECT_TRUE(alignedPtr);
-            EXPECT_FALSE((uint64_t)alignedPtr & PageSizeMask);
+            EXPECT_FALSE((size_t)alignedPtr & PageSizeMask);
 
             ReleaseRss(alignedPtr, i); // allocator can provide reused memory with RSS memory on it
             EXPECT_EQ(CountMapped(alignedPtr, i), 0ul); // no RSS memory allocated
@@ -44,21 +44,21 @@ namespace NCoro::NStack::Tests {
     TEST(StackUtilsTest, RssReleaseSeveralPages) {
         char *rawPtr, *alignedPtr = nullptr;
 
-        for (uint64_t i : {1, 2, 5, 8}) {
+        for (size_t i : {1, 2, 5, 8}) {
             EXPECT_TRUE(GetAlignedMemory(i, rawPtr, alignedPtr));
             EXPECT_TRUE(rawPtr);
             EXPECT_TRUE(alignedPtr);
-            EXPECT_FALSE((uint64_t)alignedPtr & PageSizeMask);
+            EXPECT_FALSE((size_t)alignedPtr & PageSizeMask);
 
             ReleaseRss(alignedPtr, i); // allocator can provide reused memory with RSS memory on it
             EXPECT_EQ(CountMapped(alignedPtr, i), 0ul); // no RSS memory allocated
 
-            for (uint64_t page = 0; page < i; ++page) {
+            for (size_t page = 0; page < i; ++page) {
                 *(alignedPtr + page * PageSize) = 42; // map RSS memory
                 EXPECT_EQ(CountMapped(alignedPtr, page + 1), page + 1);
             }
 
-            const uint64_t pagesToKeep = (i > 2) ? 2 : i;
+            const size_t pagesToKeep = (i > 2) ? 2 : i;
 
             ReleaseRss(alignedPtr, i - pagesToKeep);
             EXPECT_EQ(CountMapped(alignedPtr, i), pagesToKeep) << "number of pages " << i; // no RSS memory allocated
